+++ lib/command.ts
@@ -21,20 +21,32 @@ import {
 import * as print from "./print";
 
 /**
- * Try to identify the subcommand being invoked and, if there is no
- * command or the command is a reserved command, return true.
- * Otherwise return false.
+ * Determine whether sdm-local commands should be loaded.  To improve
+ * startup times and eliminate client startup when unnecessary, we do
+ * not load the sdm-local commands if we are just running a native CLI
+ * command.
  *
  * @param args command-line arguments, typically process.argv
- * @return true is the SDM local commands should be loaded
+ * @return true if the SDM local commands should be loaded
  */
-export function isReservedCommand(args: string[]): boolean {
-    const command = args.slice(2).filter(a => !/^-/.test(a)).shift();
-    if (!command) {
+export function shouldAddLocalSdmCommands(args: string[]): boolean {
+    if (args.includes("--help") || args.includes("--h") || args.includes("-h") || args.includes("-?")) {
         return true;
     }
-    const reservedCommands = ["config", "git", "gql-fetch", "gql-gen", "kube", "start"];
-    return reservedCommands.includes(command);
+    const command = args.slice(2).filter(a => !/^-/.test(a)).shift();
+    if (!command) {
+        return false;
+    }
+    const reservedCommands = [
+        "config",
+        "execute",
+        "git",
+        "gql-fetch",
+        "gql-gen",
+        "kube",
+        "start",
+    ];
+    return !reservedCommands.includes(command);
 }
 
 /**